requirements.txt
setuptools==47.3.1
tweepy==3.6.0
python-dotenv==0.9.1
PyYAML==5.1
